added multiselect

+++ server/api/data/tasks/delete.ts
@@ -3,6 +3,20 @@ import { prisma } from '../../db';
 export default defineEventHandler(async (event) => {
   const body = await readBody(event);
 
+  if (!body.clientId) {
+    throw createError({
+      status: 401,
+      statusMessage: 'Please log in again before changing the task'
+    });
+  }
+
+  if (!body.id) {
+    throw createError({
+      status: 400,
+      statusMessage: 'Please provide a task id'
+    });
+  }
+
   try {
     await prisma.tasks.delete({
       where: {
